clang-format partial

+++ defines.h
@@ -3,12 +3,12 @@
 
 typedef unsigned char u8;
 typedef unsigned short int u16;
-typedef unsigned long int u32;
-typedef unsigned long long u64;
+typedef unsigned int u32;
+typedef unsigned long int u64;
 
 typedef char s8;
 typedef short int s16;
-typedef long int s32;
-typedef long long s64;
+typedef int s32;
+typedef long int s64;
 
 #endif
